ShowMeTheColor/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
opencv-python==4.8.1.78
mediapipe==0.10.8
imutils==0.5.4
numpy==1.24.3
scikit-learn==1.3.2
matplotlib==3.7.2
colormath==3.0.0
scipy==1.11.4
pydantic==2.5.0
pillow==10.1.0